extracted MyUser into it's own app
