+++ CapstoneProject595/settings.py
@@ -42,6 +42,7 @@ INSTALLED_APPS = [
     'fitness',
     'crispy_forms',
     'bootstrap4',
+    'users',
 ]
 
 CRISPY_TEMPLATE_PACK = 'bootstrap4'
@@ -131,7 +132,7 @@ STATIC_URL = 'static/'
 DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
 
 
-AUTH_USER_MODEL = 'fitness.MyUser'
+AUTH_USER_MODEL = 'users.MyUser'
 
 LOGIN_REDIRECT_URL = 'home'
 LOGOUT_REDIRECT_URL = 'home'
